signalflow: add support for immediate computations
When immediate is set to true, the stop time will be automatically
adjusted to allow for the immediate delivery of the computation output.

This diff deprecates the current execute() method in favor of another
method that takes Long/Boolean objects instead of their primitive
equivalents.

Test Plan: compiles. wrote a small java program using the library and
verified the new flag is properly taken into account.