Upgrade to Jetty 9.x
Upgrade various dependencies of signalfx-java, most notably switching to
Jetty 9.x for important security fixes. This involved dealing with some
important jetty-websocket framework changes for the WebSocketTransport.
Verified the changes with a local test program streaming a SignalFlow
computation.

Note that we have to stick to Jetty 9.2.x as long as we intend to
support Java 7 for the signalfx-java library.